refactor: relocate ODE integrators, eigensystem solvers, and AutomaticDifferentiation into their proper namespaces
Part of a follow-up namespace-organization pass. Three independent
relocations out of the Marmot::Math grab-bag / a stray top-level root:

- explicitEuler/semiImplicitEuler/explicitEulerRichardson[WithErrorEstimator]
  move from Marmot::Math into a new Marmot::TimeIntegration header
  (MarmotTimeIntegration.h, in MarmotMechanicsCore), alongside the existing
  newmarkBetaIntegration -- same conceptual category (single-step time
  integrators), not generic math. Their tests move to a new
  TestMarmotTimeIntegration.cpp in the same module.

- computeEigenSystemJacobi/computeEigenSystemWithCardano move from
  Marmot::Math (MarmotEigenSystems.h, now deleted) into
  Marmot::FastorStandardTensors (MarmotFastorTensorBasics.h): both operate
  on FastorStandardTensors::Tensor3t/Tensor33t and already relied on
  `using namespace FastorStandardTensors;` internally -- they belong with
  the Fastor tensor family, not generic math.

- Marmot::AutomaticDifferentiation folds into
  Marmot::NumericalAlgorithms::AutomaticDifferentiation, nested next to
  the existing sibling Marmot::NumericalAlgorithms::Differentiation
  (numerical/finite-difference differentiation) -- same conceptual
  question (how to get a derivative), two methods, one location, instead
  of an unrelated top-level root.

Pure relocations, no behavior changes. Clean build and full ctest suite
(47/47, one more than before from the new TestMarmotTimeIntegration
target) verified.
