+++ modules/materials/FiniteStrainIsotropicBiotViscoelasticity/src/FiniteStrainIsotropicBiotViscoelasticity.cpp
@@ -1,7 +1,6 @@
 #include "Marmot/FiniteStrainIsotropicBiotViscoelasticity.h"
 #include "Marmot/MarmotAutomaticDifferentiationForFastor.h"
 #include "Marmot/MarmotDeformationMeasures.h"
-#include "Marmot/MarmotEigenSystems.h"
 #include "Marmot/MarmotEnergyDensityFunctions.h"
 #include "Marmot/MarmotFastorTensorBasics.h"
 #include "Marmot/MarmotFiniteStrainViscoelasticity.h"
@@ -60,7 +59,7 @@ namespace Marmot::Materials {
     const Tensor33t< scalar > C = DeformationMeasures::rightCauchyGreen( F );
 
     // compute eigenvalues and eigenvectors of C
-    auto [lam, Q] = Math::computeEigenSystemJacobi( C );
+    auto [lam, Q] = FastorStandardTensors::computeEigenSystemJacobi( C );
     Tensor33t< scalar > principalStretch( 0. );
     for ( int i = 0; i < 3; ++i ) {
       principalStretch( i, i ) = sqrt( lam( i ) );
